+++ src/Exception/InvalidBinaryException.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Darsyn\IP\Exception;
+
+class InvalidBinaryException extends InvalidIpAddressException {}